requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Werkzeug==0.16.1
waitress==2.0.0
psycopg2==2.8.6
opencv-python==3.4.8.29
numpy==1.21.2
Pillow==6.2.1
torch==1.9.1
torchvision==0.10.1
keras==2.2.4
scikit-image==0.17.2
tensorflow==1.14.0
gpsphoto==2.2.3
ExifRead==2.3.2
piexif==0.7.1
pandas==1.1.5

pip~=21.2.4
wheel~=0.37.0
attrs~=21.2.0
h5py~=3.4.0
pytz~=2021.1
scipy~=1.7.1
matplotlib~=3.4.3
PyYAML~=5.4.1
Jinja2~=3.0.1
itsdangerous~=2.0.1
click~=8.0.1
six~=1.16.0
setuptools~=58.1.0
tensorboard~=1.14.0
networkx~=2.6.3
MarkupSafe~=2.0.1
tifffile~=2021.8.30
imageio~=2.9.0
python-dateutil~=2.8.2
termcolor~=1.1.0
aniso8601~=9.0.1
jsonschema~=3.2.0
pyrsistent~=0.18.0
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.3.2